Simplified ImageJ server-related code
Removed ImagePlusServerBuilder, ImagePlusServer and BufferedImagePlusServer, putting the (few) relevant static methods into IJTools instead.
