+++ qupath-core-processing/src/main/java/qupath/imagej/detect/tissue/SimpleTissueDetection.java
@@ -46,14 +46,14 @@ import ij.plugin.filter.RankFilters;
 import ij.process.Blitter;
 import ij.process.ByteProcessor;
 import ij.process.ImageProcessor;
-import qupath.imagej.images.servers.ImagePlusServer;
-import qupath.imagej.images.servers.ImagePlusServerBuilder;
+import qupath.imagej.helpers.IJTools;
 import qupath.imagej.objects.ROIConverterIJ;
 import qupath.imagej.processing.ROILabeling;
 import qupath.imagej.processing.SimpleThresholding;
 import qupath.lib.common.GeneralTools;
 import qupath.lib.images.ImageData;
 import qupath.lib.images.PathImage;
+import qupath.lib.images.servers.ImageServer;
 import qupath.lib.objects.PathAnnotationObject;
 import qupath.lib.objects.PathObject;
 import qupath.lib.objects.PathObjects;
@@ -130,7 +130,7 @@ public class SimpleTissueDetection extends AbstractDetectionPlugin<BufferedImage
 		@Override
 		public Collection<PathObject> runDetection(final ImageData<BufferedImage> imageData, final ParameterList params, final ROI pathROI) throws IOException {
 			
-			ImagePlusServer server = ImagePlusServerBuilder.ensureImagePlusWholeSlideServer(imageData.getServer());
+			ImageServer<BufferedImage> server = imageData.getServer();
 			
 //			double downsample = getDownsampleFactor(imageData.getServer());
 			double downsample = 1;
@@ -149,7 +149,7 @@ public class SimpleTissueDetection extends AbstractDetectionPlugin<BufferedImage
 					request = RegionRequest.createInstance(server.getPath(), downsample, pathROI);
 
 				
-				pathImage = server.readImagePlusRegion(request); // TODO: Implement z-stack support
+				pathImage = IJTools.convertToImagePlus(server, request); // TODO: Implement z-stack support
 				lastPathROI = pathROI;
 			}
